requirements.txt
numpy>=1.23,<2.0
pandas>=1.5,<2.3
scipy>=1.10,<1.14
pyyaml>=6.0,<7.0
matplotlib>=3.7,<3.10
scikit-learn>=1.3,<1.6
h5py>=3.9,<3.13
tqdm>=4.66,<5.0
PyQt5>=5.15,<5.16